requirements.txt
numpy
pydicom
pynrrd
matplotlib